+++ TaskTrack/DashboardApp/TaskTrackDashboardApp.cpp
@@ -1,0 +1,6 @@
+#include "TaskTrackDashboardApp.h"
+namespace Upp {
+#include "TaskTrackDashboardAppUi.inc"
+#include "TaskTrackDashboardAppState.inc"
+#include "TaskTrackDashboardAppPanels.inc"
+}