+++ TaskTrack/DashboardCore/TaskTrackDashboardCore.cpp
@@ -1,0 +1,19 @@
+#include "TaskTrackDashboardCore.h"
+
+#include <ctype.h>
+
+namespace Upp {
+
+#include "TaskTrackDashboardCoreParseA.inc"
+#include "TaskTrackDashboardCoreParseB1.inc"
+#include "TaskTrackDashboardCoreParseB2.inc"
+#include "TaskTrackDashboardCoreModelA1.inc"
+#include "TaskTrackDashboardCoreModelA2.inc"
+#include "TaskTrackDashboardCoreModelB1.inc"
+#include "TaskTrackDashboardCoreModelB2.inc"
+#include "TaskTrackDashboardCoreStoreA1.inc"
+#include "TaskTrackDashboardCoreStoreA2.inc"
+#include "TaskTrackDashboardCoreStoreB.inc"
+#include "TaskTrackDashboardCoreSummary.inc"
+
+} // namespace Upp