+++ TaskTrack/DashboardWidgets/TaskTrackDashboardWidgets.cpp
@@ -1,0 +1,8 @@
+#include "TaskTrackDashboardWidgets.h"
+#include <math.h>
+namespace Upp {
+#include "TaskTrackDashboardWidgetsRailA.inc"
+#include "TaskTrackDashboardWidgetsRailB.inc"
+#include "TaskTrackDashboardWidgetsBody.inc"
+#include "TaskTrackDashboardWidgetsPanel.inc"
+}